fix(tui): edit_file diff popup now uses absolute work_dir instead of tildized path
Bug: edit_file popup showed plain code with line numbers instead of git
diff output (e.g. '+Duration::from_secs(5 + (n % 11))').

Root cause: run_git_diff() used workspace_dir ('~/Projects/tact') as
current_dir, but Rust's Command doesn't expand '~', so git diff failed
and the renderer fell back to inline_content + line numbers.

Fix: DiffPopup creation now uses work_dir (absolute PathBuf) instead
of workspace_dir (tildized display string).

Also removes redundant 'as u64' cast in account.rs jitter_interval().

+++ crates/tact-ui/src/account.rs
@@ -112,7 +112,7 @@ pub fn spawn_poller(account_tx: UnboundedSender<AccountUpdate>) {
 /// deterministic within the same wall-clock tick).
 fn jitter_interval() -> Duration {
     let n = POLL_COUNTER.fetch_add(1, Ordering::Relaxed);
-    Duration::from_secs(5 + (n % 11) as u64)
+    Duration::from_secs(5 + (n % 11))
 }
 
 #[cfg(test)]
